Move Offer Info to seperate component #23
